+++ src/linemod_train.cpp
@@ -41,7 +41,6 @@
 #include <ecto/ecto.hpp>
 
 #include <opencv2/core/core.hpp>
-#include <opencv2/highgui/highgui.hpp>
 #include <opencv2/imgproc/imgproc.hpp>
 #include <opencv2/objdetect/objdetect.hpp>
 
@@ -50,13 +49,12 @@
 #include <object_recognition_core/db/document.h>
 #include <object_recognition_core/db/model_utils.h>
 
-#define USE_GLUT 1
-#if USE_GLUT
-#include <object_recognition_renderer/renderer_glut.h>
-#else
-#include <object_recognition_renderer/renderer_osmesa.h>
-#endif
+#include <object_recognition_renderer/renderer3d.h>
 #include <object_recognition_renderer/utils.h>
+
+#if LINEMOD_VIZ_IMG
+  #include <opencv2/highgui/highgui.hpp>
+#endif
 
 using ecto::tendrils;
 using ecto::spore;
@@ -68,14 +66,27 @@ namespace ecto_linemod
     static void
     declare_io(const tendrils& params, tendrils& inputs, tendrils& outputs)
     {
-    inputs.declare(&Trainer::json_db_, "json_db", "The DB parameters", "{}").required(
-        true);
-    inputs.declare(&Trainer::object_id_, "object_id",
+      inputs.declare(&Trainer::json_db_, "json_db", "The DB parameters", "{}").required(true);
+      inputs.declare(&Trainer::object_id_, "object_id",
         "The object id, to associate this model with.").required(true);
+      inputs.declare(&Trainer::visualize_, "visualize", "If True, visualize the output.", true);
 
       outputs.declare(&Trainer::detector_, "detector", "The LINE-MOD detector");
       outputs.declare(&Trainer::Rs_, "Rs", "The matching rotations of the templates");
       outputs.declare(&Trainer::Ts_, "Ts", "The matching translations of the templates.");
+      outputs.declare(&Trainer::distances_, "distances", "The matching depth of the templates.");
+      outputs.declare(&Trainer::Ks_, "Ks", "The matching calibration matrices of the templates.");
+      outputs.declare(&Trainer::renderer_n_points_, "renderer_n_points", "Renderer parameter: the number of points on the sphere.");
+      outputs.declare(&Trainer::renderer_angle_step_, "renderer_angle_step", "Renderer parameter: the angle step sampling in degrees.");
+      outputs.declare(&Trainer::renderer_radius_min_, "renderer_radius_min", "Renderer parameter: the minimum scale sampling.");
+      outputs.declare(&Trainer::renderer_radius_max_, "renderer_radius_max", "Renderer parameter: the maximum scale sampling.");
+      outputs.declare(&Trainer::renderer_radius_step_, "renderer_radius_step", "Renderer parameter: the step scale sampling.");
+      outputs.declare(&Trainer::renderer_width_, "renderer_width", "Renderer parameter: the image width.");
+      outputs.declare(&Trainer::renderer_height_, "renderer_height", "Renderer parameter: the image height.");
+      outputs.declare(&Trainer::renderer_focal_length_x_, "renderer_focal_length_x", "Renderer parameter: the focal length x.");
+      outputs.declare(&Trainer::renderer_focal_length_y_, "renderer_focal_length_y", "Renderer parameter: the focal length y.");
+      outputs.declare(&Trainer::renderer_near_, "renderer_near", "Renderer parameter: near distance.");
+      outputs.declare(&Trainer::renderer_far_, "renderer_far", "Renderer parameter: far distance.");
     }
 
     void
@@ -116,59 +127,75 @@ namespace ecto_linemod
       mesh_file.close();
     }
 
-      cv::Ptr<cv::linemod::Detector> detector_ptr = cv::linemod::getDefaultLINEMOD();
-      *detector_ = *detector_ptr;
+    cv::Ptr<cv::linemod::Detector> detector_ptr = cv::linemod::getDefaultLINEMOD();
+    *detector_ = *detector_ptr;
 
-      // Define the display
-      size_t width = 640, height = 480;
-      double near = 0.1, far = 1000;
-      double focal_length_x = 525, focal_length_y = 525;
+    // Define the display
+    *renderer_width_ = 640; *renderer_height_ = 480;
+    *renderer_near_ = 0.1; *renderer_far_ = 1000.0;
+    *renderer_focal_length_x_ = 525.0; *renderer_focal_length_y_ = 525.0;
 
     // the model name can be specified on the command line.
-#if USE_GLUT
-    RendererGlut renderer = RendererGlut(mesh_path);
-#else
-    RendererOSMesa renderer = RendererOSMesa(mesh_path);
-#endif
-    renderer.set_parameters(width, height, focal_length_x, focal_length_y, near,
-        far);
+    Renderer3d renderer = Renderer3d(mesh_path);
+    renderer.set_parameters(*renderer_width_, *renderer_height_, *renderer_focal_length_x_,
+                            *renderer_focal_length_y_, *renderer_near_, *renderer_far_);
+
     std::remove(mesh_path.c_str());
 
-      RendererIterator renderer_iterator = RendererIterator(&renderer, 150);
+    *renderer_n_points_ = 150;
+    *renderer_angle_step_ = 10;
+    *renderer_radius_min_ = 0.6;
+    *renderer_radius_max_ = 1.1;
+    *renderer_radius_step_ = 0.4;
+    RendererIterator renderer_iterator = RendererIterator(&renderer, *renderer_n_points_);
+    //set the RendererIterator parameters
+    renderer_iterator.angle_step_ = *renderer_angle_step_;
+    renderer_iterator.radius_min_ = float(*renderer_radius_min_);
+    renderer_iterator.radius_max_ = float(*renderer_radius_max_);
+    renderer_iterator.radius_step_ = float(*renderer_radius_step_);
 
-      cv::Mat image, depth, mask;
-      cv::Mat_<unsigned short> depth_short;
-      cv::Matx33d R;
-      cv::Vec3d T;
-      for (size_t i = 0; !renderer_iterator.isDone(); ++i, ++renderer_iterator)
-      {
+    cv::Mat image, depth, mask;
+    cv::Matx33d R;
+    cv::Vec3d T;
+    cv::Matx33f K;
+    for (size_t i = 0; !renderer_iterator.isDone(); ++i, ++renderer_iterator)
+    {
       std::stringstream status;
       status << "Loading images " << (i+1) << "/"
           << renderer_iterator.n_templates();
       std::cout << status.str();
 
-        renderer_iterator.render(image, depth, mask);
-        R = renderer_iterator.R();
-        T = renderer_iterator.T();
+      cv::Rect rect;
+      renderer_iterator.render(image, depth, mask, rect);
 
-        depth.convertTo(depth_short, CV_16U, 1000);
+      R = renderer_iterator.R_obj();
+      T = renderer_iterator.T();
+      float distance = fabs(renderer_iterator.D_obj() - float(depth.at<ushort>(depth.rows/2.0f, depth.cols/2.0f)/1000.0f));
+      K = cv::Matx33f(float(*renderer_focal_length_x_), 0.0f, float(rect.width)/2.0f, 0.0f, float(*renderer_focal_length_y_), float(rect.height)/2.0f, 0.0f, 0.0f, 1.0f);
 
-        std::vector<cv::Mat> sources(2);
-        sources[0] = image;
-        sources[1] = depth;
+      std::vector<cv::Mat> sources(2);
+      sources[0] = image;
+      sources[1] = depth;
 
-/*      // Display the rendered image
-      cv::namedWindow("Rendering");
-      if (!image.empty()) {
-        cv::imshow("Rendering", image);
-        cv::waitKey(1);
+#if LINEMOD_VIZ_IMG
+      // Display the rendered image
+      if (*visualize_)
+      {
+        cv::namedWindow("Rendering");
+        if (!image.empty()) {
+          cv::imshow("Rendering", image);
+          cv::waitKey(1);
+        }
       }
-*/
-        detector_->addTemplate(sources, "object1", mask);
+#endif
 
-        // Also store the pose of each template
-        Rs_->push_back(cv::Mat(R));
-        Ts_->push_back(cv::Mat(T));
+      detector_->addTemplate(sources, "object1", mask);
+
+      // Also store the pose of each template
+      Rs_->push_back(cv::Mat(R));
+      Ts_->push_back(cv::Mat(T));
+      distances_->push_back(distance);
+      Ks_->push_back(cv::Mat(K));
 
       // Delete the status
       for (size_t j = 0; j < status.str().size(); ++j)
@@ -178,13 +205,28 @@ namespace ecto_linemod
       return ecto::OK;
     }
 
+    /** True or False to output debug image */
+    ecto::spore<bool> visualize_;
     /** The DB parameters as a JSON string */
     ecto::spore<std::string> json_db_;
     /** The id of the object to generate a trainer for */
     ecto::spore<std::string> object_id_;
     ecto::spore<cv::linemod::Detector> detector_;
-    spore<std::vector<cv::Mat> > Rs_;
-    spore<std::vector<cv::Mat> > Ts_;
+    ecto::spore<std::vector<cv::Mat> > Rs_;
+    ecto::spore<std::vector<cv::Mat> > Ts_;
+    ecto::spore<std::vector<float> > distances_;
+    ecto::spore<std::vector<cv::Mat> > Ks_;
+    ecto::spore<int> renderer_n_points_;
+    ecto::spore<int> renderer_angle_step_;
+    ecto::spore<double> renderer_radius_min_;
+    ecto::spore<double> renderer_radius_max_;
+    ecto::spore<double> renderer_radius_step_;
+    ecto::spore<int> renderer_width_;
+    ecto::spore<int> renderer_height_;
+    ecto::spore<double> renderer_near_;
+    ecto::spore<double> renderer_far_;
+    ecto::spore<double> renderer_focal_length_x_;
+    ecto::spore<double> renderer_focal_length_y_;
   };
 } // namespace ecto_linemod