+++ CHANGELOG.rst
@@ -1,3 +1,21 @@
+0.3.2 (2015-01-18)
+------------------
+* fix compilation on 32 bits
+* update OpenCV dependency
+* Contributors: Vincent Rabaud
+
+0.3.1 (2015-01-18)
+------------------
+* fixed object orientation,
+  fixed icp,
+  added detection of multiple objects,
+  added visualization of point clouds
+* clean extensions
+* update docs
+* Updated render function to match latest ork_renderer commit bfcdbc2c74b21ac07b20c953f994269a427eb25c
+* fix some runtime behavior
+* Contributors: Isura, Vincent Rabaud, nlyubova
+
 0.3.0 (2014-01-29  01:59:06 +0100)
 ----------------------------------
 - better mesh handling